Movement system implemented
There is a script called Movement attached to the Player GameObject. This object needs a RigidBody2D and collider2D.
The script contains a list of KeyCodes for what keys can be used in the random assignment. The random assignment happens in the ChangeKeybinds function and this function ensure the keys aren't the same.